backend/pyproject.toml
[project]
name = "loremaster-api"
version = "0.1.0"
description = "FastAPI backend for Loremaster"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115,<0.116",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.32,<0.33",
]

[tool.uv]
package = true
